backend/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
chromadb
pypdf
python-multipart
python-dotenv
langchain-google-genai
sentence-transformers
langchain-huggingface